Extract branch instructions to instructions_branch.go module
